+++ crates/bytesbuf/benches/global_pool.rs
@@ -4,8 +4,8 @@
 use std::alloc::System;
 
 use alloc_tracker::{Allocator, Session};
-use byte_sequences::{BytesView, GlobalPool};
 use bytes::BufMut;
+use bytesbuf::{BytesView, GlobalPool};
 use criterion::{BatchSize, Criterion, criterion_group, criterion_main};
 use testing_aids::repeating_incrementing_bytes;