+++ crates/bytesbuf/examples/mem_has_memory_forwarding.rs
@@ -4,8 +4,8 @@
 //! Showcases how to implement the `HasMemory` trait using the forwarding
 //! implementation strategy, whereby the memory provider of a dependency is used.
 
-use byte_sequences::{BytesBuf, BytesView, HasMemory, Memory, MemoryShared, TransparentTestMemory};
 use bytes::{Buf, BufMut};
+use bytesbuf::{BytesBuf, BytesView, HasMemory, Memory, MemoryShared, TransparentTestMemory};
 
 const PAYLOAD_LEN: usize = 12345;